Consolidated SQL Executors
Replaced the previous postgres only SQL executor with the GeneralDatabaseExecutor which currently has templates for both postgres and mysql
